requirements.txt
flask
torch
torchvision
mtcnn
numpy
opencv-python
Pillow
tqdm

